fix control for venv build

+++ control.sh
@@ -102,7 +102,9 @@ check_py() {
     fi
     source ${G_VENV_DIR}/bin/activate
     pip install -r ${G_API_DIR}/requirements.txt
-    mv ${G_VENV_DIR}/bin/uwsgi ${G_VENV_DIR}/bin/curve_uwsgi
+    if [ -e ${G_VENV_DIR}/bin/uwsgi ]; then
+        mv ${G_VENV_DIR}/bin/uwsgi ${G_VENV_DIR}/bin/curve_uwsgi
+    fi
     echo ${G_VERSION} > ${VENV_VERSION_FILE}
     deactivate
     echo "venv deployed."
@@ -148,7 +150,6 @@ check() {
     check_web
     check_py
     check_api
-    check_uwsgi
 }
 
 start() {
